Conversation Manager
Created Convesation Object data which consists of ConversationEntry class that has the initial values to store name, text, and image. I needed a management obkect that would support several lines/entries of a convo. That is added with th eimplementation of the Conversation Scriptable object and the ConversationAssetCreator.cs.

The following requires a manager class that would take in convos from characters and display them on screen. So we have initiated this manager as a singleton and imported the recommended Singleton manager class.